% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_navigation.R
\name{get_biomart_grch38_path}
\alias{get_biomart_grch38_path}
\title{get_biomart_grch38_path}
\usage{
get_biomart_grch38_path()
}
\arguments{
\item{none}{}
}
\value{
A path to the rds file.
}
\description{
Returns the path a saved mart
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
get_biomart_grch38_path
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
\seealso{
Other mart: \code{\link{get_biomart_hg38_path}},
  \code{\link{post_process_salmon}}
}
\concept{mart}
