% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_salmon.R
\name{post_process_salmon}
\alias{post_process_salmon}
\title{post_process_salmon}
\usage{
post_process_salmon(this_script_path = "", input_file_paths, output_dir,
  ref = "grch38", gene_biotypes = c("protein_coding", "IG_C_gene",
  "IG_D_gene", "IG_J_gene", "IG_V_gene", "TR_C_gene", "TR_D_gene",
  "TR_J_gene", "TR_V_gene"), output_transcript_matrix = TRUE,
  output_hgnc_matrix = TRUE, output_entrez_id_matrix = FALSE,
  output_piped_hugo_entrez_id_matrix = FALSE,
  output_upper_quartile_norm = TRUE,
  output_log2_upper_quartile_norm = TRUE, counts_or_tpm = "counts")
}
\arguments{
\item{this_script_path}{Path to script that runs this function for documentation urposes}

\item{input_file_paths}{Named character vector of paths to the pipeline output data. Samples will be named by the names of the vector. If not named thye will be named after the quant file.}

\item{output_dir}{Path to the output folder.}

\item{ref}{options 'grch38' with ensemble output or 'hg38' with ucsc output}

\item{gene_biotypes}{The type of biomaRt gene_biotypes that should be output to the gene level output.}

\item{output_transcript_matrix}{T/F}

\item{output_hgnc_matrix}{T/F}

\item{output_entrez_id_matrix}{T/F}

\item{output_piped_hugo_entrez_id_matrix}{T/F}

\item{output_upper_quartile_norm}{T/F}

\item{output_log2_upper_quartile_norm}{T/F}

\item{counts_or_tpm}{'counts' or 'tpm'}
}
\value{
A vector of paths to the files.
}
\description{
Joins individual sample files into one tsv file.  Also converts ucsc names to hgnc_symbol|entrez_ids
and outputs the gene level counts.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
post_process_salmon
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
\seealso{
Other mart: \code{\link{get_biomart_grch38_path}},
  \code{\link{get_biomart_hg38_path}}
}
\concept{mart}
