% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_star_salmon.R
\name{post_process_star_salmon}
\alias{post_process_star_salmon}
\title{post_process_star_salmon}
\usage{
post_process_star_salmon(input_file_paths, output_dir, sample_data_path,
  gene_biotypes = c("IG_C_gene", "IG_D_gene", "IG_J_gene", "IG_V_gene",
  "misc_RNA", "processed_transcript", "protein_coding", "TR_C_gene",
  "TR_D_gene", "TR_J_gene", "TR_V_gene", "lincRNA"),
  sample_folder_column = "Sample_Folder",
  sample_id_column = "Sample_ID", thread_num = 16)
}
\arguments{
\item{input_file_paths}{Character vector of paths to the pipeline output data.}

\item{output_dir}{Path to the output folder.}

\item{sample_data_path}{Path to the sample data which should contatin the sample_id_column and sample_folder_column}

\item{gene_biotypes}{The type of biomaRt gene_biotypes that should be output to the gene level output.}

\item{sample_folder_column}{The name of the column that has sample folder names}

\item{sample_id_column}{The name of the column that has sample ids}

\item{thread_num}{Integer number of threads to run mclapply statements}
}
\value{
A path to the rds file.
}
\description{
Joins individual sample files into one tsv file.  Also converts ucsc names to hgnc_symbol|entrez_ids
and outputs the gene level counts.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
post_process_star_salmon
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
\seealso{
Other mart: \code{\link{get_biomart_hsa_ucsc_path}}
}
\concept{mart}
